Dirty workaround for required properties in multipart requests
Caution: Do not use this if you don't know exactly what you're doing - because I don't and this change may break things.

We've experienced validation failures for requests with Content-Type multipart/mixed that had only one part with required keys. At the place of this change, $data is an array with the part name as it's key and the data to validate as it's value. Hence, the required key is searched for in the inner array as well.